Revert "Add MMEventHistory and integrate it into summary.rs"
This reverts commit a643cf1862b53be1f9e5012f30b5dc0fbb9a633a.

+++ mediator/src/feedback/producers/afl.rs
@@ -235,7 +235,7 @@ impl RewardFunction for AFLBranchFeedback {
         let reward = RewardEntry::new(
             SummaryProducerIdentifier::AFLBranchFeedback,
             task,
-            reward_value as f64,
+            reward_value,
         );
         log::info!(
             "[REWARD][AFLSummary] Reward: {:?} ({} newly branch hitted, {}/{} hit count increased)",
